requirements.txt
# Core
python-dotenv>=1.0.0
httpx>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Telegram
aiogram>=3.10.0

# XML parsing (stdlib, no extra dep)
